Simplifying the install process for Arch-based distros
Currently working through an install and I want to see if this functions.

I want to use wineasio from the AUR, that way the package manager is keeping track of it. Additionally, I would like to attempt to use symbolic links from the package to the Proton version so they're all in sync and tracked.

This may be unnecessary, but I hope it could make the installation easier.